added requirements

+++ backup_File/objects.py
@@ -1,6 +1,8 @@
 import os 
 import shutil 
 from tqdm import tqdm 
+import logging 
+
 
 class backupDirectory(): 
     def __init__(self, directory_txt, destination): 
@@ -12,8 +14,8 @@ class backupDirectory():
         self.destination = destination
 
     def copyDirectory(self): 
-        for dirFolder in tqdm(self.directory_list):
+        for dirFolder in tqdm(self.directory_list, position=0, leave=False):
             new_folder = os.path.split(dirFolder)[-1]
-            if os.path.isdir(dirFolder): 
+            if os.path.isdir(dirFolder) and os.path.exists(dirFolder): 
                 shutil.copytree(dirFolder, rf"{self.destination}\{new_folder}", dirs_exist_ok=True)
-                print(rf"{self.destination}\{new_folder} Sucessfully backed up")
+                #print(rf"{self.destination}\{new_folder} Sucessfully backed up")
